+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Version schema is `year.month.num_release`
 
 ## [Unreleased]
 
+### Added
+
+- Input fields for custom min/max MJD limits https://github.com/snad-space/ztf-viewer/pull/367
+- Object pages now support query parameters: `?min_mjd=58800&max_mjd=59000` https://github.com/snad-space/ztf-viewer/pull/367
+
 ### Changed
 
 - ZTF DR17 is the default DR now